[ADD] added display name, error validation message to Usuario and OfertaCategoriaModel - [ADD] implemented logged validation on actions ensuring only accessible if logged, else return the user to the log form [ADD] Error and success message to login form [ADD] success message on viewbag on signup successfull [ADD] Implemented logout controller action and functionality [ADD] Link to sign up on login form, link to login on signup form [ADD] data type to user model to validate the signup form

+++ p3p1-bolsa-empleo-new/p3p1_bolsa_trabajo_new/Controllers/OfertasController.cs
@@ -13,7 +13,7 @@ namespace p3p1_bolsa_trabajo_new.Controllers
 {
     public class OfertasController : Controller
     {
-        private p3p1BolsaTrabajoEntitiesCategoriaOferta db = new p3p1BolsaTrabajoEntitiesCategoriaOferta();
+        private p3p1BolsaTrabajoEntitiesCategoriaOfertas db = new p3p1BolsaTrabajoEntitiesCategoriaOfertas();
 
         // GET: Ofertas
         public ActionResult Index()
@@ -54,6 +54,7 @@ namespace p3p1_bolsa_trabajo_new.Controllers
             }
             else
             {
+                ViewBag.ErrorMessage = "Please Login";
                 return RedirectToAction("Login", "Login");
             }
         }
@@ -61,27 +62,43 @@ namespace p3p1_bolsa_trabajo_new.Controllers
         // GET: Ofertas/Details/5
         public ActionResult DetailsAdmin(int? id)
         {
-            if (id == null)
+            if (Session["id_usuarios"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Oferta oferta = db.Ofertas.Find(id);
+                if (oferta == null)
+                {
+                    return HttpNotFound();
+                }
+                return View("DetailsAdmin", oferta);
             }
-            Oferta oferta = db.Ofertas.Find(id);
-            if (oferta == null)
+            else
             {
-                return HttpNotFound();
+                ViewBag.ErrorMessage = "Please Login";
+                return RedirectToAction("Login", "Login");
             }
-            return View("DetailsAdmin", oferta);
         }
 
         public ActionResult ViewAll(int? category_id, int? page)
         {
-            IOrderedQueryable<Oferta> ofertas = from o in db.Ofertas where o.id_categoria_ofertas == category_id orderby o.fecha_posteo descending select o;
-            IOrderedQueryable<categoriaOfertaEmpleo> categorias = from c in db.categoriaOfertaEmpleos where c.id_categoria_ofertas == category_id orderby c.id_categoria_ofertas select c;
-            dynamic ofertas_categorias = new ExpandoObject();
-            int rows2display = 2;
-            ofertas_categorias.Ofertas = ofertas.ToPagedList(page ?? 1, rows2display);
-            ofertas_categorias.Categorias = categorias;
-            return View("ViewAll", ofertas_categorias);
+            if (Session["id_usuarios"] != null)
+            {
+                IOrderedQueryable<Oferta> ofertas = from o in db.Ofertas where o.id_categoria_ofertas == category_id orderby o.fecha_posteo descending select o;
+                IOrderedQueryable<categoriaOfertaEmpleo> categorias = from c in db.categoriaOfertaEmpleos where c.id_categoria_ofertas == category_id orderby c.id_categoria_ofertas select c;
+                dynamic ofertas_categorias = new ExpandoObject();
+                int rows2display = 2;
+                ofertas_categorias.Ofertas = ofertas.ToPagedList(page ?? 1, rows2display);
+                ofertas_categorias.Categorias = categorias;
+                return View("ViewAll", ofertas_categorias);
+            }
+            else
+            {
+                ViewBag.ErrorMessage = "Please Login";
+                return RedirectToAction("Login", "Login");
+            }
         }
 
         public ActionResult ViewAllUser(int? category_id, int? page)
@@ -123,17 +140,25 @@ namespace p3p1_bolsa_trabajo_new.Controllers
         // GET: Ofertas/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (Session["id_usuarios"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Oferta oferta = db.Ofertas.Find(id);
+                if (oferta == null)
+                {
+                    return HttpNotFound();
+                }
+                ViewBag.id_categoria_ofertas = new SelectList(db.categoriaOfertaEmpleos, "id_categoria_ofertas", "titulo", oferta.id_categoria_ofertas);
+                return View(oferta);
             }
-            Oferta oferta = db.Ofertas.Find(id);
-            if (oferta == null)
+            else
             {
-                return HttpNotFound();
+                ViewBag.ErrorMessage = "Please Login";
+                return RedirectToAction("Login", "Login");
             }
-            ViewBag.id_categoria_ofertas = new SelectList(db.categoriaOfertaEmpleos, "id_categoria_ofertas", "titulo", oferta.id_categoria_ofertas);
-            return View(oferta);
         }
 
         // POST: Ofertas/Edit/5
@@ -143,29 +168,45 @@ namespace p3p1_bolsa_trabajo_new.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "id_ofertas,titulo,descripcion,fecha_posteo,activo,id_categoria_ofertas,ubicacion,posicion,nombre_empresa")] Oferta oferta)
         {
-            if (ModelState.IsValid)
+            if (Session["id_usuarios"] != null)
             {
-                db.Entry(oferta).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("IndexAdmin");
+                if (ModelState.IsValid)
+                {
+                    db.Entry(oferta).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("IndexAdmin");
+                }
+                ViewBag.id_categoria_ofertas = new SelectList(db.categoriaOfertaEmpleos, "id_categoria_ofertas", "titulo", oferta.id_categoria_ofertas);
+                return View(oferta);
             }
-            ViewBag.id_categoria_ofertas = new SelectList(db.categoriaOfertaEmpleos, "id_categoria_ofertas", "titulo", oferta.id_categoria_ofertas);
-            return View(oferta);
+            else
+            {
+                ViewBag.ErrorMessage = "Please Login";
+                return RedirectToAction("Login", "Login");
+            }
         }
 
         // GET: Ofertas/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            if (Session["id_usuarios"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Oferta oferta = db.Ofertas.Find(id);
+                if (oferta == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(oferta);
             }
-            Oferta oferta = db.Ofertas.Find(id);
-            if (oferta == null)
+            else
             {
-                return HttpNotFound();
+                ViewBag.ErrorMessage = "Please Login";
+                return RedirectToAction("Login", "Login");
             }
-            return View(oferta);
         }
 
         // POST: Ofertas/Delete/5
@@ -173,10 +214,18 @@ namespace p3p1_bolsa_trabajo_new.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Oferta oferta = db.Ofertas.Find(id);
-            db.Ofertas.Remove(oferta);
-            db.SaveChanges();
-            return RedirectToAction("IndexAdmin");
+            if (Session["id_usuarios"] != null)
+            {
+                Oferta oferta = db.Ofertas.Find(id);
+                db.Ofertas.Remove(oferta);
+                db.SaveChanges();
+                return RedirectToAction("IndexAdmin");
+            }
+            else
+            {
+                ViewBag.ErrorMessage = "Please Login";
+                return RedirectToAction("Login", "Login");
+            }
         }
 
         protected override void Dispose(bool disposing)
